# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SpiderX is a private tracker server.",
        "title": "SpiderX API",
        "contact": {
            "name": "EndPot",
            "url": "https://github.com/endpot"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.0.1"
    },
    "host": "localhost:8888",
    "basePath": "/",
    "paths": {
...
        "/forums/{forum_id}/topics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取主题列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "获取主题列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "论坛版块ID",
                        "name": "forum_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "每页数量",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.PageResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/definitions/topic.Topic"
                                                    },
                                                    {
                                                        "type": "object",
                                                        "properties": {
                                                            "forum": {
                                                                "$ref": "#/definitions/topic.Forum"
                                                            },
                                                            "user": {
                                                                "$ref": "#/definitions/topic.User"
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "创建主题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "创建主题",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "论坛版块ID",
                        "name": "forum_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "创建主题请求参数",
                        "name": "topic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.CreateTopicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/topic.Topic"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "forum": {
                                                            "$ref": "#/definitions/topic.Forum"
                                                        },
                                                        "user": {
                                                            "$ref": "#/definitions/topic.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
        "/forums/{forum_id}/topics/{topic_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取主题信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "获取主题信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "论坛版块ID",
                        "name": "forum_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/topic.Topic"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "forum": {
                                                            "$ref": "#/definitions/topic.Forum"
                                                        },
                                                        "user": {
                                                            "$ref": "#/definitions/topic.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "删除主题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "删除主题",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "论坛版块ID",
                        "name": "forum_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "主题ID",
                        "name": "topic_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "请求成功"
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "更新主题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Topic"
                ],
                "summary": "更新主题",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "论坛版块ID",
                        "name": "forum_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "主题ID",
                        "name": "topic_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新主题请求参数",
                        "name": "topic",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topic.UpdateTopicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/topic.Topic"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "forum": {
                                                            "$ref": "#/definitions/topic.Forum"
                                                        },
                                                        "user": {
                                                            "$ref": "#/definitions/topic.User"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数异常",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "没有操作权限",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "没有对象",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ErrResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 500
                },
                "message": {
                    "description": "消息",
                    "type": "string"
                }
            }
        },
        "response.PageMeta": {
            "type": "object",
            "properties": {
                "page": {
                    "description": "页码",
                    "type": "integer",
                    "example": 1
                },
                "per_page": {
                    "description": "每页数量",
                    "type": "integer",
                    "example": 15
                },
                "total": {
                    "description": "总数",
                    "type": "integer",
                    "example": 100
                }
            }
        },
        "response.PageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                },
                "meta": {
                    "description": "附加信息",
                    "type": "object",
                    "$ref": "#/definitions/response.PageMeta"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "数据",
                    "type": "object"
                }
            }
        },
...
        "topic.CreateTopicRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "description": "主题内容",
                    "type": "string",
                    "example": "This is a content"
                },
                "position": {
                    "description": "置顶位置",
                    "type": "integer",
                    "default": 0,
                    "example": 0
                },
                "title": {
                    "description": "主题标题",
                    "type": "string",
                    "example": "This is a title"
                }
            }
        },
        "topic.Forum": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "版块ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "版块名称",
                    "type": "string",
                    "example": "This is a forum name"
                }
            }
        },
        "topic.Topic": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "主题内容",
                    "type": "string",
                    "example": "This is a content"
                },
                "created_at": {
                    "description": "创建时间戳（秒）",
                    "type": "integer",
                    "example": 1591974665
                },
                "forum": {
                    "description": "论坛版块",
                    "type": "object"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "is_hot": {
                    "description": "是否热门",
                    "type": "boolean",
                    "example": false
                },
                "position": {
                    "description": "置顶位置",
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "description": "主题标题",
                    "type": "string",
                    "example": "This is a title"
                },
                "updated_at": {
                    "description": "更新时间戳（秒）",
                    "type": "integer",
                    "example": 1591974665
                },
                "user": {
                    "description": "用户",
                    "type": "object"
                },
                "view_count": {
                    "description": "查看次数",
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "topic.UpdateTopicRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "主题内容",
                    "type": "string",
                    "example": "This is a content"
                },
                "position": {
                    "description": "置顶位置",
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "description": "主题标题",
                    "type": "string",
                    "example": "This is a title"
                }
            }
        },
        "topic.User": {
            "type": "object",
            "properties": {
                "display_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "spider"
                },
                "id": {
                    "description": "用户ID",
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}